// src/types/common.ts
export type AccountType = 'player' | 'club' | 'academy' | 'trainer' | 'agent' | 'admin' | 'marketer' | 'parent';